refactor(engine): extract JSON runtime support
