+++ crates/coral-engine/src/runtime/mod.rs
@@ -3,6 +3,7 @@
 
 pub(crate) mod catalog;
 pub(crate) mod error;
+pub(crate) mod json;
 pub(crate) mod query;
 pub(crate) mod registry;
 pub(crate) mod schema_provider;
